bpf/notify: Repurpose upper bits of version for ext_version
Currently, the highest notify capture version is 3. As such having 8
bits should be sufficient to store the version for at least the next few
decades.

We can reuse the other 8 bits for an extension version, which enables
downstream projects to extend notify messages for niche features
without wasting space in cilium/cilium.

+++ pkg/hubble/parser/threefour/parser_test.go
@@ -603,7 +603,7 @@ func TestDecodeDropNotify(t *testing.T) {
 	SrcLabel := identity.NumericIdentity(123)
 	DstLabel := identity.NumericIdentity(456)
 
-	dropNotify := func(version uint16, iptraceid ...uint64) monitor.DropNotify {
+	dropNotify := func(version uint8, iptraceid ...uint64) monitor.DropNotify {
 		var id uint64
 		if len(iptraceid) > 0 {
 			id = iptraceid[0]
